Handle datum input to jsonb construction correctly
Values which are already (non-numeric) Datums should be converted to
jsonb using the same rules that pg itself uses. In particular, json
and jsonb values should not be stringified, timestamps should be
correctly formatted, etc. Handle this by punting all such cases to
to_jsonb(). Fix a couple of type bugs in passing.