Refactor. SPI results are no longer registered tupdescs
Registering tupdescs indiscriminately turns out to be a bad idea (they
never get freed, there's a finite limit per session, and performance
may be an issue). So while we still make a typeinfo for SPI results,
we don't register it or intern it; it will be GC'd when no tuple
references it anymore.

Since an unregistered record type can't be passed to record_out
without eliciting an ugly error, provide a special __tostring for this
case to make it easy to print record values for debugging, etc. (and
also to avoid having to rewrite half the regression tests).

Avoid computing t->nested, since it requires a lot of typecache
lookups and doesn't really save us any runtime when exploding tuples.

In passing, clean up use of object type checks.