Plug a hole in error handling w.r.t. coroutines.
In order to enforce our "no Lua code when there are unhandled PG
errors" rule, we have to ensure that coroutine.resume can't end
up catching a pg error generated by the coroutine.

In passing, fix a failure to do adequate stack checking in our own
coroutine management for SRFs.

The error handling can still be defeated by a C module that calls
lua_c?pcallk? or lua_resume. This can't be helped; it's up to the
superuser not to enable trusted-language access to modules that can do
this.

(Might be worth considering an alternate strategy: allow Lua code
while there are pending errors, but not allow any interaction with PG
itself in that state. This would eliminate the need to wrap
pcall/resume/etc, but most of our functions would need to check for
pending errors which seems ugly. Also having a normal return from a
Lua function change into an error because of a pending error seems
dangerous.)