Disable the just-added luajit int8 hack by default.
It turns out that luajit cdata int8's break a bunch of otherwise
useful functionality, mostly thanks to the fact that they stringify to
values like "0LL" instead of just "0". The only actual benefit was the
ability to do arithmetic with them.

Leave the code in place behind an off-by-default compile time switch
to allow experimentation.