Improve parameter handling in numeric module.
If an argument to a numeric op isn't already a numeric datum or Lua
number, then use the pgtype.numeric constructor to get a numeric
datum. This avoids the problem where Lua will accept a long digit
string as a floating-point value with loss of precision, and also
allows numeric operands to be SQL datums castable to numerics or
luajit cdata values.

In passing fix another case of unsafe pg functions being called
outside a catch block.