Initialization tweaks.
Add new GUCs: install_globals (default on) controls whether we install
modules into the global table; on_common_init is run after the
on_trusted/untrusted_init string and runs _inside_ the sandbox unlike
the others.

Adjust behaviour of some functions so that on_init is run the same way
regardless of whether we're in postmaster or not; specifically, use
pllua_getinterpreter (which returns null before the start of phase2
init) to control most behaviour changes rather than IsUnderPostmaster.

print() is now its own module, and output goes to the log rather than
the client when running the init strings regardless of whether running
in postmaster or not.

pllua.error is now a module, mainly to have a canonical reference
place to get the pcall/lpcall function wrappers from. The table of
error names is now in the registry rather than being independently
loaded in the elog and error modules, and it uses metatables to
populate itself on first use rather than relying on callers to do
that.