Error handling rework to close holes.
Firstly, lxpcall turns out to be unsafe so stop exposing it to the
user (it still provides xpcall functionality inside on_init, so it's
not completely removed).

Secondly, use a different approach for registering errors. For one,
make sure that we keep a separate error state for each level of
activation on the stack. Second, we clean the error out of the
registry when returning back to pg proper, so it can't confuse us in a
following call or nested call. (It's important that we do this without
throwing a lua error, for which we rely on luaL_unref's guarantee of
non-throwingness.)