Ensure that post-finalization Datum values are safe.
Lua code can obtain references to post-finalization objects via the
keys of ephemeron tables. Most of our objects are refobjects where we
set the pointer to null on finalization, but Datums weren't doing this
and could give garbage results or crash on post-finalization access.
Avoid the problem by nulling out the metatable in the finalizer,
leaving the object as an inert and opaque blob.