Cruft and dependency cleanup
Move pg-version and lua-version cruft to separate include files.

Use Lua instead of Perl for build-time generation of errcodes and
functable headers.

Disable pushcfunction workaround in Lua 5.4 as unnecessary. (It'll
also be unnecessary in 5.3.5, but don't touch that yet, because of the
issue of handling a 5.3.5 compile linked against a 5.3.4 shlib.)