Return type / datum construction changes/fixes.
Array construction can now take a single table (or collection) and
turn it into a 1-d array sized based on the largest integer key.

Row construction from a single nil or no args at all works regardless
of the arity of the rowtype.

In passing standardize the "for k,v in pairs(x)" idiom to use in C
code via helper functions.