Fix modification of row datums.
Brown-paper-bag time: when modifying a row datum that was already
modified, the code to explode the row bailed out before actually
pushing the deform table onto the lua stack, causing the subsequent
assignment to fail (and by coincidence, silently).

Move the check for already-exploded rows to _after_ getting the deform
table, and add a whole pile of regression tests to give the explosion
code more of a workout.

Fixes #5.